build: resolve conflicting types
Conflicts between `unsigned int` and size_t in definition and
declaration for several lookup_ functions.

Solves the following type of errors for gcc 7.2.0:

In file included from src/basic/cap-list.c:31:0:
src/basic/cap-from-name.h:141:1: error: conflicting types for ‘lookup_capability’
 lookup_capability (register const char *str, register size_t len)
 ^~~~~~~~~~~~~~~~~
src/basic/cap-list.c:28:38: note: previous declaration of ‘lookup_capability’ was here
 static const struct capability_name* lookup_capability(register const char *str, register unsigned int len);

+++ src/basic/af-list.c
@@ -25,7 +25,7 @@
 #include "util.h"
 #include "af-list.h"
 
-static const struct af_name* lookup_af(register const char *str, register unsigned int len);
+static const struct af_name* lookup_af(register const char *str, register size_t len);
 
 #include "af-to-name.h"
 #include "af-from-name.h"
