+++ src/basic/errno-list.c
@@ -25,7 +25,7 @@
 #include "errno-list.h"
 
 static const struct errno_name* lookup_errno(register const char *str,
-                                                 register unsigned int len);
+                                                 register size_t len);
 
 #include "errno-to-name.h"
 #include "errno-from-name.h"
